requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
pymongo==4.5.0
python-dotenv==1.0.0
email-validator==2.0.0
bcrypt==4.0.1
PyJWT==2.8.0
Werkzeug==2.3.7
gunicorn==21.2.0
redis==5.0.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-mongoengine==0.31.0
mongoengine==0.27.0
validators==0.22.0 